feat : 처음 두장 카드 합 블랙잭인 경우 로직 추가

+++ src/main/kotlin/blackjack/dto/PlayerInfo.kt
@@ -2,5 +2,5 @@ package blackjack.dto
 
 data class PlayerInfo(
     val name: String,
-    val betAmount: Int
+    val betAmount: Double
 )
